requirements.txt
langchain==0.0.350
langchain-community==0.0.10
python-dotenv==1.0.0
faiss-cpu==1.7.4
pypdf2==3.0.0
sentence-transformers==2.2.2
huggingface-hub==0.16.4
transformers==4.30.0
numpy==1.24.0
deepseek-api==0.1.0